+++ config/file-browser.php
@@ -1,6 +1,0 @@
-<?php
-
-// config for Mydnic/FilamentFileBrowser
-return [
-
-];